requirements.txt
# TadqeeqAI v3.0 Requirements
# Core dependencies for the RAG system

# Core RAG
chromadb>=0.4.0
sentence-transformers>=2.2.0
rank-bm25>=0.2.2
ollama>=0.1.0

# Document Processing
PyMuPDF>=1.23.0
python-docx>=0.8.11
reportlab>=4.0.0

# Desktop UI
pywebview>=4.0.0

# Utilities
numpy>=1.24.0
tqdm>=4.65.0

# Optional for development
pyinstaller>=5.0.0
